The app can now use the real database from the server. It is defaulted to use the stub DB and can be changed in Main.java in the startup method. Just uncomment ServerAccessDB and comment out ServerAccessStub.
Added a new class: ServerAccessDB to interact with the server. I also
had to update the other server access classes to add an extra open()
method to work with the server thread.

The tests for FetchQuestion are disabled right now. I had to update the
JSON parser amd haven't fixed the tests yet. I also got rid of that ugly
string and replaced it with a nice one.